Moved constrained specs back to cardano-ledger-test
