--- eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Constrained.hs
@@ -1,14 +1,0 @@
-module Test.Cardano.Ledger.Conway.Constrained (
-  module X,
-)
-where
-
-import Test.Cardano.Ledger.Conway.Constrained.Instances as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Cert as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Deleg as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Gov as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.GovCert as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Ledger as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.PParams as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Pool as X
-import Test.Cardano.Ledger.Conway.Constrained.Spec.Utxo as X
