+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance.hs
@@ -8,10 +8,10 @@
 {-# LANGUAGE UndecidableInstances #-}
 
 module Test.Cardano.Ledger.Conformance (
-  module Test.Cardano.Ledger.Conformance.ExecutableSpecRule.Core,
+  module Test.Cardano.Ledger.Conformance.ExecSpecRule.Core,
   module Test.Cardano.Ledger.Conformance.SpecTranslate.Core,
 ) where
 
-import Test.Cardano.Ledger.Conformance.ExecutableSpecRule.Core
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Core
 import Test.Cardano.Ledger.Conformance.Orphans ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Core
