+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/Spec/Conway.hs
@@ -5,8 +5,8 @@ module Test.Cardano.Ledger.Conformance.Spec.Conway (spec) where
 
 import Cardano.Ledger.Conway (Conway)
 import Test.Cardano.Ledger.Conformance (conformsToImpl)
-import Test.Cardano.Ledger.Conformance.ExecutableSpecRule.Conway ()
-import Test.Cardano.Ledger.Conway.Constrained (ConwayFn)
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Conway ()
+import Test.Cardano.Ledger.Constrained.Conway
 import Test.Cardano.Ledger.Conway.ImpTest (withImpState)
 import Test.Cardano.Ledger.Imp.Common
