+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Gov.hs
@@ -7,7 +7,7 @@
 
 -- | Specs necessary to generate, environment, state, and signal
 -- for the GOV rule
-module Test.Cardano.Ledger.Conway.Constrained.Spec.Gov where
+module Test.Cardano.Ledger.Constrained.Conway.Gov where
 
 import Data.Foldable
 
@@ -26,8 +26,8 @@ import Constrained
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Crypto (StandardCrypto)
-import Test.Cardano.Ledger.Conway.Constrained.Instances
-import Test.Cardano.Ledger.Conway.Constrained.Spec.PParams
+import Test.Cardano.Ledger.Constrained.Conway.Instances
+import Test.Cardano.Ledger.Constrained.Conway.PParams
 
 govEnvSpec ::
   IsConwayUniv fn =>
